+++ pkg/trace/agent/agent.go
@@ -2,6 +2,7 @@ package agent
 
 import (
 	"context"
+	"github.com/StackVista/stackstate-agent/pkg/features"
 	"github.com/StackVista/stackstate-agent/pkg/trace/interpreter"
 	"sync/atomic"
 	"time"
@@ -55,6 +56,9 @@ type Agent struct {
 
 	// Used to synchronize on a clean exit
 	ctx context.Context
+
+	// Features Supported by the StackState Backend
+	SupportedFeatures *features.Features
 }
 
 // NewAgent returns a new Agent object, ready to be started. It takes a context
@@ -88,6 +92,7 @@ func NewAgent(ctx context.Context, conf *config.AgentConfig) *Agent {
 	sw := writer.NewStatsWriter(conf, statsChan)
 	svcW := writer.NewServiceWriter(conf, filteredServiceChan)
 	sie := interpreter.NewSpanInterpreterEngine(conf)
+	sf := features.NewFeatures(conf)
 
 	return &Agent{
 		Receiver:              r,
@@ -104,6 +109,7 @@ func NewAgent(ctx context.Context, conf *config.AgentConfig) *Agent {
 		ServiceExtractor:      se,
 		ServiceMapper:         sm,
 		SpanInterpreterEngine: sie,
+		SupportedFeatures: 	   sf,
 		obfuscator:            obf,
 		tracePkgChan:          tracePkgChan,
 		conf:                  conf,
@@ -119,11 +125,6 @@ func (a *Agent) Run() {
 	// itself too fast (nightmare loop)
 	watchdogTicker := time.NewTicker(a.conf.WatchdogInterval)
 	defer watchdogTicker.Stop()
-	////
-	//featuresTicker := time.NewTicker(5 * time.Second)
-	//
-	//// Channel to announce new features detected
-	//featuresCh := make(chan features.Features, 1)
 
 	// update the data served by expvar so that we don't expose a 0 sample rate
 	info.UpdatePreSampler(*a.Receiver.PreSampler.Stats())
@@ -140,6 +141,7 @@ func (a *Agent) Run() {
 	a.ErrorsScoreSampler.Run()
 	a.PrioritySampler.Run()
 	a.EventProcessor.Start()
+	a.SupportedFeatures.Start()
 
 	for {
 		select {
@@ -161,6 +163,7 @@ func (a *Agent) Run() {
 			a.ErrorsScoreSampler.Stop()
 			a.PrioritySampler.Stop()
 			a.EventProcessor.Stop()
+			a.SupportedFeatures.Stop()
 			return
 		}
 	}
@@ -210,7 +213,9 @@ func (a *Agent) Process(t pb.Trace) {
 	for _, span := range t {
 		a.obfuscator.Obfuscate(span)
 		Truncate(span)
-		a.SpanInterpreterEngine.Interpret(span)
+		if a.SupportedFeatures.FeatureEnabled("interpreted-spans") {
+			a.SpanInterpreterEngine.Interpret(span)
+		}
 	}
 	a.Replacer.Replace(&t)
